+++ src/Core/LogColorizer.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <QString>
+
+class LogColorizer {
+public:
+    static QString formatToHtml(const QString &rawLog);
+};